LPS-93516 Move integration test UserPermissionTest of com.liferay.portal.service.permission package to module
